fix: SearchTest
